backend/requirements.txt
boto3
fastapi
langchain
langchain_community
langchain_core
langchain_google_genai
langgraph
matplotlib
numpy
pandas
pdfminer.six
pdfplumber
pinecone
pydantic
pydantic_settings
PyPDF2
python-dotenv
Requests
snowflake_connector_python
snowflake_snowpark_python
uvicorn
sentence_transformers
